refactor: some changes based on PR feedback
