+++ src/ditto/api.py
@@ -34,7 +34,6 @@ from .configs import (
     TRTLLMModelConfig,
     TRTLLMOptimizationProfileConfig,
     TRTLLMPluginConfig,
-    TRTLLMQuantConfig,
 )
 from .constants import INPUT_IDS
 from .convert import convert
@@ -143,7 +142,7 @@ def trtllm_build(
             transformed_graph_module,
             argument_hint,
             build_config=TRTLLMBuildConfig.merge(profile_config, model_config),
-            quant_config=TRTLLMQuantConfig.create_from(global_quant_config) if global_quant_config else None,
+            global_quant_config=global_quant_config,
             trt_config=trt_config,
             engine_cache=engine_cache,
             network_name=get_network_name(model),
@@ -206,7 +205,7 @@ def build_trtllm_engine_components(
     argument_hint: TRTLLMArgumentHint,
     *,
     build_config: TRTLLMBuildConfig,
-    quant_config: TRTLLMQuantConfig | None = None,
+    global_quant_config: GlobalQuantConfig | None = None,
     trt_config: TensorRTConfig | None = None,
     engine_cache: BaseEngineCache | None = None,
     network_name: str | None = None,
@@ -219,7 +218,7 @@ def build_trtllm_engine_components(
         graph_module (GraphModule): The graph module to convert to TensorRT engines
         argument_hint (TRTLLMArgumentHint): Configuration for input arguments
         build_config (TRTLLMBuildConfig): Configuration for building TensorRT-LLM engines
-        quant_config (TRTLLMQuantConfig | None, optional): Quantization configuration. Defaults to None.
+        global_quant_config (GlobalQuantConfig | None, optional): Global quantization configuration. Defaults to None.
         trt_config (TensorRTConfig | None, optional): TensorRT builder configuration. Defaults to None.
         engine_cache (BaseEngineCache | None, optional): Cache for storing/loading built engines. Defaults to None.
         network_name (str | None, optional): Name of the network. Defaults to None.
@@ -238,7 +237,7 @@ def build_trtllm_engine_components(
                 graph_module,
                 build_config,
                 argument_hint.mapping,
-                quant_config=quant_config,
+                global_quant_config=global_quant_config,
                 architecture=network_name,
             ),
         )
