+++ src/ditto/quantization.py
@@ -21,7 +21,6 @@ from compressed_tensors.config import CompressionFormat
 from compressed_tensors.quantization import QuantizationScheme, QuantizationStrategy
 from loguru import logger
 from tensorrt_llm.quantization import QuantAlgo
-from tensorrt_llm.quantization.functional import unpack_int32_into_int8
 from torch._ops import OpOverload
 from transformers import PretrainedConfig, PreTrainedModel
 from transformers.utils.quantization_config import (
@@ -283,131 +282,3 @@ def resolve_qlinear_device_map(model: PreTrainedModel) -> None:
         if isinstance(module, qlinear_cuda_old.QuantLinear):
             if module.wf.device != module.qzeros.device:
                 module.wf = module.wf.to(module.qzeros.device)
-
-
-def unpack_qweight(qweight: torch.Tensor, bits: int, quant_method: QuantizationMethod) -> torch.Tensor:
-    """Unpack the quantized weight tensor.
-
-    Args:
-        qweight (torch.Tensor): The quantized weight tensor
-        bits (int): The number of bits used for quantization
-        quant_method (QuantizationMethod): The quantization method used
-
-    Returns:
-        torch.Tensor: The unpacked weight tensor
-    """
-    if quant_method in (QuantizationMethod.GPTQ, QuantizationMethod.AWQ):
-        assert bits in (4, 8), f"Unsupported GPTQ or AWQ bits: {bits=}"
-        if bits == 4:
-            qweight = unpack_int32_into_int8(
-                qweight if quant_method is QuantizationMethod.AWQ else qweight.T,
-                quant_method is QuantizationMethod.AWQ,
-            )
-            qweight = qweight.T if quant_method is QuantizationMethod.GPTQ else qweight
-            qweight = qweight - 8
-            qweight[qweight < 0] += 16
-            qweight = qweight.view(torch.uint8).contiguous()
-        else:
-            qweight = (
-                qweight.T.contiguous().view(torch.uint8).T.contiguous()
-                if quant_method is QuantizationMethod.GPTQ
-                else qweight.view(torch.uint8).contiguous()
-            )
-            qweight = (qweight - 128).to(torch.int8)
-    elif quant_method == QuantizationMethod.COMPRESSED_TENSORS:
-        assert bits == 8, f"Unsupported bits: {bits=}. Only 8-bit quantization (FP8) is supported currently."
-        assert qweight.dtype.itemsize == 1, f"Wrong dtype of qweight: {qweight.dtype=}"
-    else:
-        raise NotImplementedError(f"Unsupported quantization method: {quant_method}")
-
-    return qweight
-
-
-def unpack_zeros(zeros: torch.Tensor, bits: int, quant_method: QuantizationMethod) -> torch.Tensor:
-    """Unpack the quantized zero point tensor.
-
-    Args:
-        zeros (torch.Tensor): The quantized zero point tensor
-        bits (int): The number of bits used for quantization
-        quant_method (QuantizationMethod): The quantization method used
-
-    Returns:
-        torch.Tensor: The unpacked zero point tensor
-    """
-    if quant_method in (QuantizationMethod.GPTQ, QuantizationMethod.AWQ):
-        assert bits in (4, 8), f"Unsupported GPTQ or AWQ bits: {bits=}"
-        if bits == 4:
-            zeros = unpack_int32_into_int8(zeros, quant_method is QuantizationMethod.AWQ)
-        else:
-            zeros = zeros.view(torch.int8)
-        zeros = -zeros + 2 ** (bits - 1) - 1 * (quant_method is QuantizationMethod.GPTQ)
-    elif quant_method == QuantizationMethod.COMPRESSED_TENSORS:
-        assert bits == 8, f"Unsupported bits: {bits=}. Only 8-bit quantization (FP8) is supported currently."
-        assert zeros.dtype.itemsize == 1, f"Wrong dtype of zeros: {zeros.dtype=}"
-    else:
-        raise NotImplementedError(f"Unsupported quantization method: {quant_method}")
-
-    return zeros
-
-
-def postprocess_qweight_for_trtllm(
-    qweight: torch.Tensor,
-    bits: int,
-    quant_method: QuantizationMethod,
-    *,
-    model_dtype: torch.dtype,
-) -> torch.Tensor:
-    """Postprocess the quantized weight tensor for TensorRT-LLM.
-
-    Args:
-        qweight (torch.Tensor): The quantized weight tensor
-        bits (int): The number of bits used for quantization
-        quant_method (QuantizationMethod): The quantization method used
-        model_dtype (torch.dtype): The model data type.
-
-    Returns:
-        torch.Tensor: The postprocessed weight tensor for TensorRT-LLM
-    """
-    if quant_method in (QuantizationMethod.GPTQ, QuantizationMethod.AWQ):
-        assert qweight.dtype in (torch.uint8, torch.int8), f"Unsupported tensor dtype: {qweight.dtype=}"
-        assert bits in (4, 8), f"Unsupported GPTQ or AWQ bits: {bits=}"
-        if bits == 4:
-            qweight = (qweight[:, 1::2] * 16 + qweight[:, ::2]).view(torch.int8)
-        weight_dtype = torch.int8 if bits == 8 else torch.quint4x2
-        qweight = torch.ops.trtllm.preprocess_weights_for_mixed_gemm(qweight, weight_dtype, torch.float16).view(
-            model_dtype
-        )
-    else:
-        raise NotImplementedError(f"Unsupported quantization method: {quant_method}")
-
-    return qweight
-
-
-def postprocess_zeros_for_trtllm(
-    zeros: torch.Tensor,
-    bits: int,
-    quant_method: QuantizationMethod,
-    *,
-    scale: torch.Tensor,
-    model_dtype: torch.dtype,
-) -> torch.Tensor:
-    """Postprocess the quantized zero point tensor for TensorRT-LLM.
-
-    Args:
-        zeros (torch.Tensor): The quantized zero point tensor
-        bits (int): The number of bits used for quantization
-        quant_method (QuantizationMethod): The quantization method used
-        scale (torch.Tensor): The scale tensor
-        model_dtype (torch.dtype): The model data type
-
-    Returns:
-        torch.Tensor: The postprocessed zero point tensor for TensorRT-LLM
-    """
-    if quant_method in (QuantizationMethod.GPTQ, QuantizationMethod.AWQ):
-        assert bits in (4, 8), f"Unsupported GPTQ or AWQ bits: {bits=}"
-        zeros = zeros * scale
-        zeros = zeros.to(model_dtype)
-    else:
-        raise NotImplementedError(f"Unsupported quantization method: {quant_method}")
-
-    return zeros
